manage user

+++ index.js
@@ -3,7 +3,7 @@ const app = express();
 const port = 3000;
 const path = require("path");
 const router = express.Router();
-var { exec } = require("child_process");
+var { exec, spawn } = require("child_process");
 
 app.use(express.json());
 app.use(express.urlencoded());
@@ -13,16 +13,17 @@ app.get("/", (req, res) => {
 
 app.post("/run", (req, res) => {
   const { imageName } = req.body;
-  let containerName;
-  if (imageName === "ubuntu:18.04") {
-    containerName = "ubuntu";
-  }
+  const containers = {
+    "ubuntu:18.04": "ubuntu",
+    "hello-world": "hello",
+    centos: "cent",
+  };
   exec(
-    `sudo docker run --name ${containerName} -t -d ${imageName}`,
+    `sudo docker run --name ${containers[imageName]} -i -d ${imageName}`,
     (error, stdout, stderr) => {
-      if (stdout) {
-        console.log("stdout", stdout);
-      }
+      if (error) return res.json(error);
+      if (stdout) return res.json(stdout);
+      if (stderr) return res.json(stderr);
     }
   );
   return;
@@ -33,10 +34,14 @@ app.post("/stop", (req, res) => {
   exec(
     `sudo docker container stop ${containerName}`,
     (error, stdout, stderr) => {
+      if (error) return res.json(error);
+      if (stderr) return res.json(stderr);
       if (stdout) {
         exec(`sudo docker rm ${containerName}`, (error, stdout, stderr) => {
+          if (error) return res.json(error);
+          if (stderr) return res.json(stderr);
           if (stdout) {
-            return;
+            return res.json(stdout);
           }
         });
       }
@@ -45,7 +50,7 @@ app.post("/stop", (req, res) => {
 });
 
 app.get("/images", (req, res) => {
-  exec("sudo docker image ls", (err, stdout, stderr) => {
+  exec("docker image ls", (err, stdout, stderr) => {
     return res.json(stdout);
   });
 });
@@ -55,17 +60,27 @@ app.get("/containers", (req, res) => {
     return res.json(stdout);
   });
 });
+
+app.post("/createuser", (req, res) => {
+  const { username, password } = req.body;
+  exec(`mkpasswd ${password}`, (err, stdout, stderr) => {
+    if (stdout) {
+      const hashpw = stdout.split("\n")[0];
+      exec(`sudo useradd -m -p ${hashpw} -s /bin/bash ${username}`, () => {
+        return res.json(username);
+      });
+    }
+  });
+});
+
+app.post("/deleteuser", (req, res) => {
+  const { username } = req.body;
+  exec(`sudo deluser --remove-home ${username}`, (err, stdout, stderr) => {
+    if (stdout) {
+      return res.json(stdout)
+    }
+  });
+});
 app.listen(port, () => {
   console.log(`Example app listening at http://localhost:${port}`);
 });
-
-function createDocker() {
-  exec("sudo docker run --name hello1 hello-world", (error, stdout, stderr) => {
-    return stdout;
-  });
-}
-function runDocker() {
-  exec("sudo docker container start -i hello1", (error, stdout, stderr) => {
-    return stdout + "";
-  });
-}
